Fix mac deployment issues (#44). (#82)
* Make sure local dev keeps working after deployment (#44).

Our deployment packages zips for every platform. Currently, we ensure
that the last platform is Linux to match the App Engine platform, which
means on that other platforms we end up with dependencies for the wrong
platform after deployment.

To fix this we install dependencies again if we are not on Linux.

Also, fix Go App Engine deployment.