Don't mark build as bad when leaks are detected.
Leaks can be frequent, however they don't impact reproducing
of regular crashes (as they are reported during shutdown). So,
instead of marking build as bad and blocking fuzzing, we instead
let the build go through which helps with regular fuzzing and
then report the leak with a proper testcase report.