Calculate new_features and new_edges metrics only after INITED.
This happened after
https://github.com/google/clusterfuzz/commit/2f0b0ab5c6efa5c0614ccea8ce63f0f9264ff89f
when merge_new_features metric was removed and we started calculating
new_features and new_edges metrics from start and end of fuzzing run.
The bug was that we were setting edge_coverage and feature_coverage
metrics before INITED or while seed corpus is processed. So, if we
crashed right after INITED, we will have a negative value for
new_features and new_edges.