Computing impacted Chromium builds for components.
The intent of this change is to fill in the Security_Impact
field for Clusterfuzz bugs which are reported in a component
(typically, V8). The task of looking up V8 branch points
and applying these labels takes a significant amount of manual
Sheriffing time so it's desirable to automate it. Ensuring
these labels are present results in various downstream flows
'just working', especially regarding release-blocking and
merge requests.

Bug 534360 has previously wanted to add builds for stable
and beta in order to calculate this information. The present
change instead does this using revision numbers alone, using
the regression range calculated already by Clusterfuzz and
the 'branched-from' text which can be retrieved from a
revision description.

Specifically this change:

1. Spots when we're trying to deduce impacts of a component
   bug (as opposed to a Chromium bug)
2. Looks up the component revisions associated with the
   Chromium revision for beta and stable (using pre-existing
   code)
3. Fetches the commit description for that particular
   component revision
4. Tries to extract a 'Cr-Branched-From:' comment from the
   revision description.
5. Compares that with the regression range in which the bug
   was found and applies Security_Impact labels (again,
   using pre-existing code)

To make this work a few minor tweaklets were required
elsewhere:

1. Parts of blame_task_test.py were abstracted into a
   superclass.
2. impact_task_test.py had various fake version numbers
   amended to match the data which backed blame_task_test.py.
3. A small tweak to the 'which' test so it works on my
   gLinux environment.

Bug: 534360
Change-Id: I610a320540549d2d5c312b685f513d0c133e9651