Add randomized fuzzer selection to Fuchsia runs.
Apologies for the length of the CL; I promise it's shorter than it looks
(though I can chop it up further).

Also apologies for the delay; there was a Fuchsia hack-it week, and some
instability/thrashing on the Fuchsia side of things.

This PR does the following:

* Adds the fuchsia_utils directory, which contains a (hopefully) stable
interface for working with Fuchsia. This includes logic for extracting
Fuzzers from a Fuchsia build.

* Moves the Fuchsia build logic into build_manager.  While it's a very
tiny class right now, and doesn't take much advantage of build_manager
features, this should allow it to e.g. use the weighted random selection
in future CLs, handle revision # logic more gracefully, etc.

* Adds the ability to *choose* a fuzzer at random, and run it.  (Left
for a future CL is extracting the crash log from the device and placing
it someplace useful, though *writing* the log should be handled
automagically by our wrapper class.)

My only question is the license headers: I added the standard Google LLC
license header to the files we're yoinking from Fuchsia, but *also* left
the Fuchsia license header in place.  I do not know whether we can just
merge these, if there's an elegant way to handle it, etc--I would've
just left the files totally intact, but the linter complained about the
header not looking right :)  So yeah, let me know.