Allow testcase in group to be associated to existing issue.
Previously, we narrowed the scope of existing issue updates
to similar testcases not existing in group. This made things
those testcases in group to be visible only in UI (as part
of group), but didn't help developers assigned to issue. So,
expand scope to update bug directly if a better reproducible
testcase is available and current attached one became
uninteresting.