[fuchsia] Enables reproducible crashes for Fuchsia testcases.
To detect crashes from reproducer runs, we need to (1) properly return
the return code from the fuzz run, and (2) check for some text that only
occurs in reproducer runs.  This CL tweaks some of the code in
util/fuchsia to propertly return this information, and tweaks the way
that code is called, accordingly.

In addition, the way Fuchsia dumps logs by default means that the
stacktrace occurs *before* the `==ERROR: .*` line, which confuses the
stack analyzer, so we tweak the log processing subroutine to drop that
line at the beginning.

Finally, we re-add in some code that was accidentally removed, which
suppresses stdout during fuzzing runs (quiet=True), since Clusterfuzz
can prematurely halt jobs that appear to have crashed, before we've
finished processing all the Fuchsia logs.

And, finally-finally, there's an actual implementation of
run_single_testcase in the Fuchsia runner here :)