Remove symbolize=1 from LSAN_OPTIONS.
Defer to symbolize from ASAN_OPTIONS.

This was breaking our de-duplication because it caused inline frames to
be included in the crash state. A better long term fix in the future
will be to ignore inline frames in stack_analyzer instead, after which
we can turn this back on and get suppressions working again.