Fix regression with AFL variant task not working. (#1083)
After new engine impl was enabled in testcase_manager's
test_for_crash_with_retries, it uses the testcase.get_fuzz_target
which picks up original fuzzing engine and not the one for the
variant task. For those cases, use a cloned testcase with the
right fuzzer_name and overridden_fuzzer_name so that it sets up
AFL engine fuzzing properly.