[fuchsia] Update regex for finding relevant logs. (#1157)
Something in one of our fuzzers is matching on '==([0-9]*)==' with
'==()=='.  Unfortunately, this match tries to use the value of [0-9]* to
process logs from a specific PID, and thus fails because the captured
value is ''.

We can fix this by converting the regex to find [0-9]+ instead.

If this becomes a recurring issue, perhaps we can update Fuchsia to emit
logs in a more guaranteed-unique structure?