Speed up cross pollinate fuzzer list operation. (#1173)
* Speed up cross pollinate fuzzer list operation.

Previously we would query a job entity for every fuzz target. Now it only queries jobs for the selected targets and jobs. This has slight downside of getting an empty list when jobs don't have backup bucket, but in most cases, we would have one. Also, use the default backup bucket from project.yaml when available, otherwise we wont pollinate for cases when a default one is used.