Replace App Engine NDB with google cloud NDB. (#1298)
Part of migration to Python 3. This is our last major blocker and
dependency on the Python 2 App Engine SDK.

Also remove ndb_patcher, as that is no longer required.

Notable changes:
- indexed=False is no longer supported for StringProperty. Use TextProperty instead, which is
  equivalent.
- batch_size query keyword arg is no longer supported (Datastore backend change).
- fetch_page() is no longer supported for complex queries with !=, OR etc.
- Many import hacks/workarounds to make things work for App Engine.
  - Notable: add `grpcio` to app.yaml, cron-service.yaml
- Add workaround for googleapis/python-ndb#292 and googleapis/python-ndb#295
- Threads may be created by the new NDB implementation, which are only guaranteed to be 
  stopped after a GC. GAE doesn't like it when there are stray threads at the end of a request.
- Stub out `ndb_utils.get_all_from_query`. This is no longer required, and callers will be cleaned up in a future PR.
- All datastore operations now require:
```python
from datastore import ndb_init
with ndb_init.context():
  ...
```
This is automatically set up in tests, as well as in run_bot, run, heartbeat, run_heartbeat, and butler run, as well as for each individual request.