[fuchsia] Drop sanitizer suffixes from Zircon fuzzer names (#1745)
Due to build changes on the Fuchsia side, Zircon fuzzers no longer end
in e.g. '.asan', so we need to stop appending it unless required for
backwards compatibility with older builds. Prior to this fix, newer
builds were failing to run Zircon fuzzers because the name was now
wrong.