Request bisection for regression earlier. (#1919)
Previously, a regressed and fixed bisect were requested at the same time
when a Testcase is detected to be fixed in progression_task.

This PR changes this: A regressed bisection can be requested earlier,
even if the Testcase isn't fixed. To ensure that tasks are only
requested once we add corresponding testcase metadata.