Ensure validation and generation work always.
If no checkpoint has been saved, validation and data generation will
fail as they imply loading saved weights to use in the rebatched model.

A better design would be to have callbacks that do validation and
generation and all TensorBoard stuff but for that we first need to
convert to a dataset representation and do automatic batching.