Add FoundIn labels on Monorail.
The FoundIn=NN label is a standard for marking which milestone
in which functional bugs are found within Monorail.
ClusterFuzz should adopt this standard for the following reasons:

1) Why not? ClusterFuzz has this information so it would be great
   to add it.
2) Sheriffbot currently relies on Security_Impact when deciding on
   merge decisions. Currently, Security_Impacts are 'promoted'
   when a new build goes to stable or to beta. This sometimes is
   problematic when a single milestone is both stable and beta.
3) More generally, switching the security sheriffing procedures
   from Security_Impact to FoundIn would be less confusing.

A future change would ideally also set RegressedIn, but that's
a more significant project.

For more information internally within Google, see
go/clusterfuzz-foundin

Change-Id: Id6c89bc80922137bb82ab22f01b4df922b21f58c