[fuchsia] Work around bug in older builds with newer KVM
Newer KVM versions exposed a bug that caused Fuchsia to fail to boot.
This was previously fixed on the Fuchsia side, but requires a
ClusterFuzz-side workaround in order to be able to run builds from
before that fix rolled (e.g. when bisecting).

This commit also adds the `invtsc` feature to the QEMU CPU config to
bring it in line with Fuchsia defaults and improve performance, and
makes it slightly easier to reproduce the QEMU commands run by
ClusterFuzz.