Add support for parsing LKL style crashes (#2088)
* Add support for parsing LKL style crashes

LKL or Linux Kernel Library allows libfuzzer to be used to fuzz the
linux kernel by including it as a userspace library.  Crashes appear as
a libfuzzer "deadly signal" with a KASAN crash inside.  This change
allows for parsing these KASAN crashes with higher priority than the
libfuzzer crash to insure the correct crash_type and state.