Move timeout subtraction outside Engine.fuzz.
Currently it's not obvious at all to users of Engine.fuzz how long
fuzzing will go for when a max_time is passed.

Add Engine.fuzz_additional_processing_timeout so that the caller can
explicitly subtract the necessary processing time themselves if needed.

Remove engine_common.POSTPROCESSING_TIME. This was a holdout from the
launcher days. This is moved to the AFL launcher and can be removed
there as well once that's migrated to the Engine interface.