project_setup: clean up jobs/projects at the end of all setup sources.
Previously this clean up was happening per project setup source, leading
to potential races with a job getting deleted in one source and getting
re-added in a later source..

Another functional change for simplicity here is to not set up jobs for
disabled projects. This is no longer required as now we auto-close bugs
with invalid jobs.