Prevent potential redundant issue comment with uploaded testcases.
When a new testcase is uploaded and processed, the cleanup task will
update the testcase's issue with the crash description.

However it's possible that a new testcase is uploaded which results in a
new issue being filed. If cleanup task runs after the filing, it means
we get a redundant crash description comment added to the bug.

Prevent this by recording the bug_information at time of upload and only
updating the issue if it matches at time of running cleanup.