add EMULATED_ANDROID platform (#2267)
* add EMULATED_ANDROID platform

This allows downloading an emulator bundle (from the
ANDROID_EMULATOR_BUCKET_PATH variable) and running it (using the
ANDROID_EMULATOR_SCRIPT_PATH variable) before fuzzing.

The emulated Android instance is controlled using 'adb', just as a
physical Android device would. The new EmulatedAndroidLibFuzzerRunner
shares most logic with the existing AndroidLibFuzzerRunner.

* fix import order

* fix formatting using 'butler.py format'

* fix lint errors

* resolve first round of feedback

* change execute_command() to run_command() for 'adb devices' call

* move emulator startup to build_manager.py

This also scans the stdout of the emulator process to get the expected
device serial, which is determined at runtime. This allows us to set the
ANDROID_SERIAL environment variable.

* fix test errors from unused imports

* wait for emulated device to come online

* bail out early from functions when using emulated Android device

* add case to wait_until_fully_booted() for emulators

* fix lint error - unused import

* clean up platform check, check for correct env var