Detect Jazzer exceptions without message
Java exception do not have to have a message (e.g. StackOverflowError)
and if they don't, the first line of the Jazzer crash report will not
contain two colons and thus not be detected as a crash.

Since the "== Java Exception:" marker should be unique enough on its
own, this commit makes the Jazzer regex look only for this prefix and
not for an additional colon.