Stop any stale emulators before starting a new one. (#2285)
* change emulator to write to build_dir, not $BOT_TMPDIR

* stop any stale emulator instances before starting a new one

Since different emulators may need to be stopped different ways, we'll
depend on a 'stop' script in the bundle. In most cases, this will
contain a call to 'killall'.

* use run_and_wait() instead of run()

Testing didn't catch this at first because most scripts are very quick,
but in theory it could take longer. We want to wait until the script
finishes before continuing on.