Find CHECK failures also without source file
The block of code to replace a generic 'Fatal error' by the more
specific 'CHECK failure' was only executed if we had a source file for
the fatal error. In d8 / chrome release builds, we won't have a source
file though, but we still want to extract the CHECK failure.