Update Testcase timestamp when analyze_task completes. (#2323)
Triage task waits data_types.MIN_ELAPSED_TIME_SINCE_REPORT for a
testcase to make sure it is considered for any grouping before we filed
a bug. This is based on the `timestamp` attribute.

However, for uploaded testcases, the timestamp was previously set to the
time of upload, rather than the time when analyze_task completes and
marks the Testcase as "Processed". This means these testcases could get
filed without being grouped.

Also increase MIN_ELAPSED_TIME_SINCE_REPORT by an hour to be safe, as
the grouper can be quite slow with lots of open testcases.