Set FoundIn for impact-head bugs.
Previously ClusterFuzz recorded FoundIn labels for Chromium
bugs which impacted stable or beta. If a bug was found to impact
only Head, no FoundIn label was recorded.

With this change, we now add a FoundIn label for such situations
which makes for easier downstream triage and handling of such
security bugs.

Most of this consists of recording an extra build number in the
Impacts section of each test case.

The Impact task explicitly checks for reproducibility against
Stable and Beta builds. This change doesn't alter anything
there - we don't test against any additional builds because
ClusterFuzz always already reproduces things against the
head revision.

Instead, we just record the known dev/canary build within the
Impact structure such that we can later rely on it when filing
the issue.

Bug: https://crbug.com/1222484