Impact: mark latest milestone if code too recent.
Previously, ClusterFuzz sometimes bisected regression revisions to
versions more recent than *any* available build.

For Chromium builds, this resulted in issue_filer.py being unable
to apply a FoundIn label to the bugs it created, since the test was
not actually found in any specific build. This was completely
reasonable behavior, but was unhelpful for downstream Chromium
processes such as merging, noting regressions, crediting
reporters etc.

With this change, ClusterFuzz records a 'head' impact for the
overall milestone (e.g. '50') even if it can't be sure that
it impacts a specific build (e.g. '50.1.2.3'). This will enable
ClusterFuzz to add suitable FoundIn labels to bugs it files in
this circumstance.

Bug: https://crbug.com/1230588
Change-Id: Ib63e963fe5a4e24e548240fe2b7c06a02545dd5c