[fuchsia] Enforce timeouts on undercoat calls
The Fuchsia libFuzzer runner was ignoring the timeouts passed to it,
meaning that a bot could get stuck for a long time if the fuzzing run
hangs for any reason. As a fallback mechanism in this sort of situation,
this plumbs the timeout through to the call to the subprocess, adding a
significant amount of padding along the way (for reasons described in
the comments).