[Jira bug] Use jira issue object to fix ticket creation bug (#2797)
Current implementation:
Whenever we want to create a new jira ticket we first create a blank
ticket(with dummy values) and then update that ticket with desired
fields.

Issues:
If the update to these tickets fail (eg: wrong field/value) Then we end
up create a lot of dummy tickets because of retries.

Solution:
Eliminate the step to create a ticket with dummy values. Looking at how
things are implemented for monorail the right approach here would be to
use jira issue objects locally to store all the information and create
the ticket with all the desired values with the help of the object. In
case if ticket creation fails then this does not create garbage tickets