Deprecated references and behaviors associated with ANDROID_EMULATOR. (#2858)
ANDROID_EMULATOR was an enum originally created to enable fuzzing of
TrustyOS on Android emulators. This is no longer needed as Trusty
fuzzing has moved to on-device, and so this PR aims to clean up code
specific to these now unmaintained emulators.

The enum ANDROID_EMULATOR, however, remains associated and in use with
Trusty fuzzers and is not removed here (it may need to be renamed, after
considering all config files, in a later PR).