Fix remove file on hang afl device (#2894)
When AFL++ fuzz run is initializing and running through the corpus
inputs, if one of those corpus files hangs then ClusterFuzz needs to
remove that file from the input. It tries to remove this file from the
host machine for Android, but needs to remove the file from the Android
device instead. This PR is to update that.