Remove calculation of impact on production builds.
ClusterFuzz previously had two mechanisms to calculate impact on
production builds:

* Plan A: attempt the test case on builds corresponding to stable, beta
  and dev channels.
* Plan B: if no such builds were available, estimate based on bisection
  ranges and branch points.

Plan A stopped working properly because we stopped getting a feed of
production builds, for unknown reasons. Additionally, it was on shaky
ground because those channels imprecisely matched the numeric release
channels (sometimes there's a whole branch in between 'beta' and 'dev',
for example).

With this PR we remove all the code related to Plan A and switch to Plan
B.

Bug: https://crbug.com/1383117