[consolidation] Handle errors in a structured way. (#3070)
Previously, error handling in tasks (such as, retrying a task or closing
a testcase), could be done immediately when the error was detected in
code. With the rearchitecture, this is no longer possible.
To replace this create a structured method of handling errors, then
handle those errors using the new method.
So far only setup_testcase is partially migrated.