[unittest] Fix minijail unit test by upgrading the pyfakefs version.
When running the unit tests locally, test_minijail_env_vars fails with
an IsADirectoryError (Is a directory in the fake filesystem: '/tmp').

It looks as if this is an issue with pyfakefs misbehaving on certain
platforms in newer versions of Python: https://github.com/pytest-dev/pyfakefs/issues/686

By updating the pipfile to a newer version of pyfakefs, the unit test
now passes.