Set up filters for utasks on pub/sub (#3212)
Start setting up a system to allow linux workers to execute the
preprocess and postprocess parts of utasks.
The first and last part of a utask can be executed on any platform since
it only involves running the ClusterFuzz code which is cross-platform
and not the fuzzers which are native code and platform-specific.
Therefore Linux workers will need to read *specific* task commands from
non-linux queues.
This PR helps with that end by setting up filters to allow linux workers
to read just preprocess and postprocess commands from non-linux queues.
It also will be used by non-linux workers to ignore preprocess and
postprocess commands from those same queues.
A notable change is that we need to seperate our definition of the type
of each task from the execution (commands.py) of that task. This is
because many tasks depend on the task scheduling module, which now needs
to know the types of tasks to filter. Thus we seperate the types to
avoid circular imports.