Allow remote postprocessing of analyze task based on config. (#3352)
Add support for remote postprocessing.
More specifically, this will allow the postprocess step of analyze_task
to execute on another bot, instead of the bot executing analyze_task.
This will be necessary to support use of untrusted workers (which are
not being used yet).
Allow remote postprocessing for analyze task on an opt-in basis.