Add initial python file with boilerplate for issue_tracker implementation
This is the first commit to add Google's issue_tracker as an option for filing and updating bugs found by ClusterFuzz.

It seems to make sense to carry forward the pattern that we see with Jira and Monorail, of having an __init__ file that contains Issue, Action, and IssueTracker classes, backed by an IssueTrackerManager class in a separate file, handling i/o.

This commit starts small, creating a new directory for issue_tracker and including only __init__ with stubs for each of those classes we expect to need.