Use Extension Fields from Config in Issue Filer
In a previous iteration, we used an existing 'label' value to trigger
the Chrome Security workflow for security issues. In our new approach,
we are looking to two 'extension fields' to be set under 'security' on
the config for google-issue-tracker: ext_collaborators and
ext_issue_access_limit.

In a subsequent PR, we will add logic for the security workflow,
depending on these extension fields from the config.