Fix serialization.
1. Don't use UworkerMsg's underlying proto as the source of truth before
serializing. To type check, we copy any messages immediately to the proto.
This can lead to errors as the UworkerMsg can still be used and assigned
to, leaving proto's representation of the UworkerMsg stale. To deal with
this: Re-set all values on proto fields before serializing (and do the same
to any submessages so that we properly handle (great) grandchildren).
2. Handle assignment of list values.