Don't index jobs for Fuzzer entities. (#3619)
Some of our deployments are hitting index limits, so we can't do this
anymore.

Generally this shouldn't impact performance very much, because there are
not too many Fuzzer entities in all of our deployments.

For the existing usages of this index, just loop through and filter out
the Fuzzer entities that match.

Three usages are impacted:

- Jobs page: Tested and verified that this doesn't impact performance
much (due to pagination).
- update_mappings_for_job: Whenever a Fuzzer/Job association is
modified. Will monitor the performance of our project setup cron jobs.
- external_users: This only impacts fuzzer stats, testcase upload, which
are both one-off loads that shouldn't impact performance much.

For #3618.