Ensure that uworker_env only has string keys
Otherwise we will get errors when adding it to to a proto.
This bug pattern is very likely in ClusterFuzz because ClusterFuzz's
env vars don't need to be strings (this seems like the wrong decision IMO),
but their representation in proto must be strings since they must be
a single type.

This fixes an error introduced here: #3536
This fixes this bug:
https://pantheon.corp.google.com/errors/detail/CPCzn83Hxtr1ag;service=;version=;filter=%5B%22%5C%22_collections_abc.py%5C%22%22%5D?e=-13802955&mods=logs_tg_prod&project=google.com:cluster-fuzz