[Monorail->IssueTracker] Add Rollback Capability
This update adds roll_back and batch_size arguments to the
monorail/issuetracker migration script.

We need feedback on how best to run this script. It seems that we are
intended to run it with butler.py, but it's not clear whether butler
supports additional arguments, and it seems to require arguments (like
script_name) that we are unsure where we get values for. Thank you for
your review, and please advise!