[analyze][rearchitecture] Reduce delays on untrusted parts of analyze task
Instead of preprocessing and then putting on the queue for scheduling,
schedule right away. This eliminates up to 1 minute of time spent
pulling tasks from the queue and can improve execution speed on batch
by:
1. Causing batch to schedule the job earlier because the job
requires less CPUs than jobs with many tasks. These CPUs may not
be available in times of high-usage.
2. Allowing us to set a higher priority for the job.