Do not overwrite OOM with generic "Abrt" (#3649)
In V8 fuzzers with ASan enabled we see crashes where after a detected
out-of-memory we see an ASan "Abrt" report which overwrites the
out-of-memory state. This is not helpful, we should classify such
crashes as out-of-memory.

This also adds some more tests using existing test data but setting the
`REPORT_OOMS_AND_HANGS` environment variable.