[reland] archive: rework archive classes to make it extendable (#3515) (#3714)
In chrome, we have very specific needs when it comes to unzipping. For
instance, every chrome fuzzers are package with a `.runtime_deps` file
that lists all the runtime dependencies needed by the target. We could
use this to selectively unpack our archives.

Similarly, we could use ripunzip to improve our zipping process.

This rework will enable those new features more easily, but also make
the code clearer on what exactly is going on.

---------

Co-authored-by: jonathanmetzman <31354670+jonathanmetzman@users.noreply.github.com>