Add options to log to stackdriver from non appengine/k8s bots (#3641)
On some of our android fuzzing hosts, we don't have the fluentd server
that takes care of propagating the python logs to stackdriver. I suggest
this change that adds the option to log there directly. It was a bit of
a mess to add the additional fields given that we are stuck with a
relatively old version of the google cloud logging api. I also am in a
situation were we have three GCP projects involved (one for the service
account, one for the queues and one for the logging).
Also I added a bunch of variables to make it customizable (To chose
which to activate in the console/file/fluentd/gcp logs). Alternatively,
we might be better of using an input dictconfig or fileconfig
(https://docs.python.org/3/library/logging.config.html).