Handle fuzz target metadata correctly and in an untrusted manner. (#3894)
Previously analyze task was not retrieving the fuzz target metadata
properly (this needs the build, analyze was doing it in postprocess) and
progression was doing it in an unprivileged way (using the db in
utask_main). Change this so we can retrieve the metadata in utask_main
and save it in postprocess.
Leave a TODO in corpus_pruning which isn't migrated yet.