Fix query when throttling issue filing per project. (#3914)
"and" is not the right way to combine query conditions. "X and Y" is Y
and will only filter on Y.
Add a test to verify query correctness.
This fixes a bug preventing issue filing for some of our users.

Also log throttling as an error. It's a pretty serious step that
deserves attention if we have to throttle issue filing.