Update severity field on existing Google IssueTracker issues (#3919)
This seems to have never worked.

It's unclear to me whether we should support the label being removed as
well or not. I can't find out whether the severity field is translated
to a severity label when reading back from the tracker, in which case
edits to remove the label should probably be reflected in the severity
field. Otherwise, I think we can only add the label.

Also adds an error log in the case where there are too many severity
labels, so we can notice and diagnose such issues.

Doing this revealed a subtle bug in the existing custom handling code
for other "fake" labels, like OS. Label values are found with a
case-sensitive match in `_extract_label()` or `_get_labels()`, but
removed with a case-insensitive match in
`LabelStore.remove_by_prefix()`. This should probably be fixed, but I
did not want to change existing behavior beyond the scope of this fix.

Fixes https://crbug.com/332396351.