Delete regular build before symbolization. (#3932)
Alternative attempt to fix issues identified in
https://crbug.com/333014970, as suggested by @alhijazi on #3925.

After the bot downloads a regular build and uses it to try reproducing
the crash with larger ASAN redzone sizes, this change makes the bot
delete said regular build. It will then download a symbolized release
build and (optionally) a symbolized debug build.

This is still a bit silly in the case where the regular build and
symbolized release build are one and the same: we download and unpack
the same (potentially-large) zip twice. However, it is much less complex
and risky than #3925, which is a good tradeoff.