Shorten some paths.
ClusterFuzz currently passes an unusually long TMPDIR variable into
child processes when it's merging coverage. Chromium assumes a
reasonably short TMPDIR because it passes it into the UNIX domain socket
path within a sockaddr_t, and thus Chromium fails to launch when in this
mode.

After some investigation as described on
https://issues.chromium.org/issues/342197914
it appears as though the less risky solution is to change ClusterFuzz.
We could switch to radically different way of figuring out temporary
directory paths, but this approach in this PR is just to chop off some
characters where it's simple and easy to do so, which should unblock
Chromium fuzzers for now.