Adding CLUSTERFUZZ_RELEASE env var to batch docker container (#4210)
### Motivation

The last move to upgrade our fleet to 3.11 is switch batch. This PR adds
the CLUSTERFUZZ_RELEASE env var to batch config, so we can roll
forward/back between 3.7 and 3.11 on batch by maneuvering between the
appropriate docker image and code release

This PR belongs to this
[initiative](https://github.com/google/clusterfuzz/issues/4059)

### Deployment strategy

This code should be deployed to uworker instances, which are the actual
ones that create batch jobs. Deploy this using --release=candidate, with
the appropriate clusterfuzz-config