Fix FileExistsError (#4255)
There seems to be a race condition where the existence of a directory is
checked and if it doesn't exist it is created. But somehow this
directory is getting created in between these two steps. Maybe it is
because of multiple processes running on hosts. Replace check with
exist_ok=True.

Fixes:
https://pantheon.corp.google.com/errors/detail/CO_5v-P9kpbcoQE;locations=global?e=-13802955&mods=logs_tg_prod&project=clusterfuzz-external