Instrumenting cronjob exit codes (#4270)
### Motivation

Kubernetes signals that cronjobs fail, after retries, through
[events](https://kubernetes.io/docs/reference/kubernetes-api/cluster-resources/event-v1/).

GKE does not make it easy to alert on this failure mode. For cronjobs
without retries, the failure will be evident in the GCP panel. For
cronjobs with retries, there will be a success indicator for the last
successful cronjob, and the failure will be registered under the events
panel.

### Alternatives considered

Options available to monitor failing cronjobs are:

- the container/restart_count metric, from
[GKE](https://cloud.google.com/monitoring/api/metrics_kubernetes). This
will be flaky, since a job might succeed on the third attempt. Also, it
is not easy to pinpoint the cronjob, since we get the container name as
a label
- alert on a log based metric from the cluster events. The output of
kubectl get events gets dumped to cloud logging, so we can create a
metric on events of the form "Saw completed job:
oss-fuzz-apply-ccs-28786460, status: Failed". However this requires
regex manipulation, has to be manually applied across all projects, and
also makes it hard to derive the failing cronjob from the container
name. It also adds a hard dependency on kubernetes.

### Solution

The proposed solution is to reuse the builtin clusterfuzz metrics
implementation, and add a gauge metric CLUSTERFUZZ_CRON_EXIT_CODE, with
the cron name as a label.

If the metric is at 1, then the cronjob is unhealthy, otherwise it is
healthy. An alert must be set for when it reaches the 1 state, for every
label.
 
Since cronjobs are ephemeral, there is no need for a thread to
continuously flush metrics. The option to use monitoring without a
flushing thread was added. The same approach can be used to fix metrics
for swarming/batch.
 
Also, the flusher thread was changed to make sure that leftover metrics
are flushed before it stops.
 
Note: this PR is part of this
[initiative](https://github.com/google/clusterfuzz/issues/4271)