[build_manager] add support for remote zip
This adds support for remote ZIP.

As of now, performances are quite good locally, and the read ahead mechanism
should keep reasonable performance. Also, given that the ClusterFuzz bots are
having HDD, numbers might even be better there, as we're only storing on disk
when unpacking the build.

The memory consumption of this new feature is contant: it uses at most (and
most of the time) 50 MB of RAM.