Fix autohealing policy update logic. (#4264)
There are two logic fixes:

1. API field names are in `camelCase`, not `snake_case`.
2. Previously, we would treat trying to update a single field as an
error and delete the policy.
It is only an error to *supply* a single field in the config, but we may
legitimately find a single field needs updating.

Also moved config validation to loading time. It would likely be best to
explode if the auto-healing policy in the config has a single field set,
instead of ignoring it with a warning, but this preserves behavior for
now.