Revert Kubernetes cronjob instrumentation (#4293)
As far as Kubernetes cronjobs go, successful terminations can be
obtained through log base metrics in GCP, with the following query:

```
jsonPayload.involvedObject.name =~ "${cronjob_name}"
jsonPayload.message = "Job completed"
resource.type="k8s_cluster"
```

This metric suffices to alert on, since the abscence of successful runs
for 24h suffices to signal that a cron is broken. It is obtained from
kubernetes events (kubectl get events), so we can rely on the Kubernetes
control plane for cronjob instrumentation instead of rolling our own.

Thus, this PR will revert the previous metrics added to Kubernetes
cronjobs, for the sake of simplifying the codebase.